Additional class to read in stops.txt file.
This file has a new readBusStations class, which is designed to read in stops.txt file and create two arrays:
first one is linesArray which simply stores each line of text file.
second one is stationsArray which is a 2D array and stores both stop id and stop name.
This whole class is made so my future TST one can work. Rn I don't know wether it will be even useful or not.