Update for TST class.
I have adopted code from the Princeton book. I also wrote some code in the main: it fills the tries and also takes the string input. As an output, it gives all bus stops names which starts or equal to the string input. As it only gives the bus stop name, and not the full information about the bus stop, class is not yet finished. 
I will also soon fix the issue with the repeating code (line 212 - 219). 
For this class to work, I also had to implement two other classes StdIn and Queue (also taken from the book).