Support classes for TST.java
These two classes are needed so the TST class can function. Both these classes were taken from the Princeton and barely edited. It also should be mentioned that not all code in these two classes is necessarily for the TST to function, so I will probably delete some of the unnecessary code after we discuss it with the group during our next discord meeting.